requirements.txt
annotated-types==0.7.0
ansible==8.7.0
ansible-core==2.15.12
ansible-runner==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
baddns==1.1.798
bbot==1.1.8
bcrypt==4.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudcheck==3.1.0.318
colorama==0.4.6
cryptography==42.0.8
deepdiff==7.0.1
dnspython==2.6.1
docutils==0.21.2
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
lockfile==0.12.2
lxml==5.2.2
MarkupSafe==2.1.5
omegaconf==2.3.0
ordered-set==4.1.0
packaging==24.1
pexpect==4.9.0
pluggy==1.5.0
psutil==5.9.8
ptyprocess==0.7.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.3.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-whois==0.9.4
PyYAML==6.0.1
Random-Word==1.0.11
requests==2.32.3
requests-file==2.1.0
resolvelib==1.0.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
tabulate==0.8.10
tldextract==3.5.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
websockets==12.0
wordninja==2.0.0
xmltodict==0.12.0
xmltojson==2.0.2
